pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scider"
version = "0.1.0"
description = "Scientific Data-centric End-to-end Researcher"
keywords = ["scientific research", "data-centric", "end-to-end", "AI assistant", "autonomous research"]
authors = [
    { name = "Ke Lin", email = "leonard.keilin@gmail.com" },
    { name = "Yilin Lu", email = "lu000661@umn.edu" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "claude-agent-sdk>=0.1.33",
    "ddgs>=9.6.1",
    "epam-indigo==1.35.0",
    "feedparser>=6.0.12",
    "filetype>=1.2.0",
    "jinja2>=3.1.6",
    "json-repair>=0.53.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.2",
    "litellm>=1.79.0,<1.80.0",
    "loguru>=0.7.3",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pexpect>=4.9.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pyfunctional>=1.5.0",
    "python-toon>=0.1.2",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "scikit-learn>=1.8.0",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.10",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.9.0",
    "torchvision",
]
cu128 = [
    "torch>=2.9.0",
    "torchvision",
]
mac = [
    "torch>=2.9.0",
    "torchvision",
]
openhands = [
    "openhands-sdk==1.3.0",
    "openhands-tools==1.3.0",
]

[tool.setuptools]
packages = { find = { include = ["scider", "scider.*"] } }

[tool.setuptools.package-data]
"scider" = ["**/*.yaml"]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cu128" },
        { extra = "mac" },
    ],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-mac", extra = "mac" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-mac", extra = "mac" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-mac"
url = "https://pypi.org/simple"
explicit = true

[tool.uv.workspace]
members = [
    "benchmarks/datascibench",
    "benchmarks/scicodebench",
]
